Modify `my-count-lines-window' to count the number of logical lines.
Up to now, `my-count-lines-window' counts the number of physical lines.
So if there arefolding lines, `my-count-lines-window' counts a wrong number.
Counting logical lines provides this problem.

But a new problem appears. Because `my-count-lines-window' uses length of
a line as to count logical lines, `my-count-lines-window' counts a wrong
number if a buffer uses japanese character.