Settings about anything.
- Add migemo to anything source
To add migemo to anything source enable to search candidates by migemo.

- for filelist
Methods to list file by anything are `anything-for-files' and
`anything-filelist+'. `anything-for-files' use locate to search files, so it
can't search well in windows. `anything-filelist+' is extension of this
method.  It uses filelist made in advance. However, it doesn't work well why
I don't know.  Meanwhile, I use `anything-for-files'.

- `anything-show-kill-ring'
Show all the contents of `kill-ring' and select one by anything.

- anything-orgcard.el
anything-orgcard enables to search functions used in `org-mode'.  When I use
in key-chord bind, it cannot load in startup. Its dependency is difficult.